update: table investasi fund

+++ database/migrations/2020_07_07_133203_create_investasi_funding_table.php
@@ -19,8 +19,8 @@ class CreateInvestasiFundingTable extends Migration
             $table->bigInteger('investee_id')->nullable()->unsigned();
             $table->string('status', 1)->default('0');
             $table->string('status_tempo', 1)->default('0');
-            $table->integer('donasi_target');
-            $table->integer('donasi_masuk');
+            $table->decimal('donasi_target', 20, 2);
+            $table->decimal('donasi_masuk', 20, 2);
             $table->string('bank');
             $table->string('no_rekening');
             $table->string('atas_nama');
